+++ internal/org/org.go
@@ -5,7 +5,6 @@ import (
 	"crypto/rand"
 	"encoding/base64"
 	"errors"
-	"fmt"
 	"time"
 
 	"github.com/alexedwards/argon2id"
@@ -21,8 +20,8 @@ import (
 type Role string
 
 const (
-	Admin  Role = "admin"
-	Member Role = "member"
+	AdminRole  Role = "admin"
+	MemberRole Role = "member"
 )
 
 var (
@@ -74,21 +73,6 @@ type Org interface {
 		internalHandle string,
 		password string,
 	) (*identity.Identity, error)
-	AuthenticateMember(ctx context.Context, handle string, password string) (bool, error)
-}
-
-// Store is the registry of all orgs on a pear instance.
-// It routes DIDs to their org and provides cross-org membership checks.
-type Store interface {
-	GetOrg(ctx context.Context, orgID string) (Org, error)
-	GetOrgForDID(ctx context.Context, did syntax.DID) (Org, error)
-	AuthenticateMember(ctx context.Context, handle string, password string) (bool, error)
-	CreateOrg(
-		ctx context.Context,
-		name string,
-		adminHandle string,
-		adminPassword string,
-	) (orgID string, id *identity.Identity, err error)
 }
 
 type inviteTokenClaims struct {
@@ -111,7 +95,7 @@ func NewOrg(
 	db *gorm.DB,
 	signingSecret []byte,
 ) (*orgImpl, error) {
-	if err := db.AutoMigrate(&member{}, &spentToken{}); err != nil {
+	if err := db.AutoMigrate(&Member{}, &spentToken{}); err != nil {
 		return nil, err
 	}
 	return &orgImpl{
@@ -131,9 +115,9 @@ func (s *orgImpl) LoginMethod() string {
 }
 
 func (s *orgImpl) AddAdmin(ctx context.Context, admin syntax.DID) error {
-	result := s.db.WithContext(ctx).Model(&member{}).
+	result := s.db.WithContext(ctx).Model(&Member{}).
 		Where("org_id = ? AND member = ?", s.orgID, admin.String()).
-		Update("role", Admin)
+		Update("role", AdminRole)
 	if result.Error != nil {
 		return result.Error
 	}
@@ -156,10 +140,10 @@ func (s *orgImpl) addMemberTx(
 	return tx.WithContext(ctx).Clauses(clause.OnConflict{
 		Columns:   []clause.Column{{Name: "org_id"}, {Name: "member"}},
 		DoNothing: true,
-	}).Create(&member{
+	}).Create(&Member{
 		OrgID:        s.orgID,
 		Member:       did.String(),
-		Role:         string(Member),
+		Role:         string(MemberRole),
 		PasswordHash: passwordHash,
 		CreatedAt:    time.Now(),
 	}).Error
@@ -173,9 +157,9 @@ func (s *store) bootstrapAdmin(ctx context.Context, bootstrapSecret string, admi
 */
 
 func (s *orgImpl) GetAdmins(ctx context.Context) ([]syntax.DID, error) {
-	var rows []member
+	var rows []Member
 	if err := s.db.WithContext(ctx).
-		Where("org_id = ? AND role = ?", s.orgID, Admin).
+		Where("org_id = ? AND role = ?", s.orgID, AdminRole).
 		Find(&rows).
 		Error; err != nil {
 		return nil, err
@@ -192,7 +176,7 @@ func (s *orgImpl) GetAdmins(ctx context.Context) ([]syntax.DID, error) {
 }
 
 func (s *orgImpl) GetMembers(ctx context.Context) ([]syntax.DID, error) {
-	var rows []member
+	var rows []Member
 	if err := s.db.WithContext(ctx).Where("org_id = ?", s.orgID).Find(&rows).Error; err != nil {
 		return nil, err
 	}
@@ -210,8 +194,8 @@ func (s *orgImpl) GetMembers(ctx context.Context) ([]syntax.DID, error) {
 func (s *orgImpl) DowngradeAdmin(ctx context.Context, admin syntax.DID) error {
 	var adminCount int64
 	if err := s.db.WithContext(ctx).
-		Model(&member{}).
-		Where("org_id = ? AND role = ?", s.orgID, Admin).
+		Model(&Member{}).
+		Where("org_id = ? AND role = ?", s.orgID, AdminRole).
 		Count(&adminCount).
 		Error; err != nil {
 		return err
@@ -220,17 +204,17 @@ func (s *orgImpl) DowngradeAdmin(ctx context.Context, admin syntax.DID) error {
 		return ErrLastAdmin
 	}
 	return s.db.WithContext(ctx).
-		Model(&member{}).
-		Where("org_id = ? AND member = ? AND role = ?", s.orgID, admin.String(), Admin).
-		Update("role", Member).
+		Model(&Member{}).
+		Where("org_id = ? AND member = ? AND role = ?", s.orgID, admin.String(), AdminRole).
+		Update("role", MemberRole).
 		Error
 }
 
 func (s *orgImpl) RemoveAdmin(ctx context.Context, admin syntax.DID) error {
 	var adminCount int64
 	if err := s.db.WithContext(ctx).
-		Model(&member{}).
-		Where("org_id = ? AND role = ?", s.orgID, Admin).
+		Model(&Member{}).
+		Where("org_id = ? AND role = ?", s.orgID, AdminRole).
 		Count(&adminCount).
 		Error; err != nil {
 		return err
@@ -239,8 +223,8 @@ func (s *orgImpl) RemoveAdmin(ctx context.Context, admin syntax.DID) error {
 		return ErrLastAdmin
 	}
 	return s.db.WithContext(ctx).
-		Where("org_id = ? AND member = ? AND role = ?", s.orgID, admin.String(), Admin).
-		Delete(&member{}).
+		Where("org_id = ? AND member = ? AND role = ?", s.orgID, admin.String(), AdminRole).
+		Delete(&Member{}).
 		Error
 }
 
@@ -250,15 +234,15 @@ func (s *orgImpl) RemoveMembers(ctx context.Context, members []syntax.DID) error
 		dids = append(dids, did.String())
 	}
 	return s.db.WithContext(ctx).
-		Where("org_id = ? AND member IN ? AND role = ?", s.orgID, dids, Member).
-		Delete(&member{}).
+		Where("org_id = ? AND member IN ? AND role = ?", s.orgID, dids, MemberRole).
+		Delete(&Member{}).
 		Error
 }
 
 func (s *orgImpl) IsAdmin(ctx context.Context, did syntax.DID) (bool, error) {
-	var row member
+	var row Member
 	err := s.db.WithContext(ctx).
-		Where("org_id = ? AND member = ? AND role = ?", s.orgID, did.String(), Admin).
+		Where("org_id = ? AND member = ? AND role = ?", s.orgID, did.String(), AdminRole).
 		First(&row).
 		Error
 	if errors.Is(err, gorm.ErrRecordNotFound) {
@@ -268,7 +252,7 @@ func (s *orgImpl) IsAdmin(ctx context.Context, did syntax.DID) (bool, error) {
 }
 
 func (s *orgImpl) IsMember(ctx context.Context, did syntax.DID) (bool, error) {
-	var row member
+	var row Member
 	err := s.db.WithContext(ctx).
 		Where("org_id = ? AND member = ?", s.orgID, did.String()).
 		First(&row).
@@ -387,7 +371,7 @@ func (s *orgImpl) AuthenticateMember(
 		return false, nil
 	}
 
-	var row member
+	var row Member
 	err = s.db.WithContext(ctx).
 		Where("org_id = ? AND member = ?", s.orgID, id.DID.String()).
 		First(&row).
@@ -405,162 +389,4 @@ func (s *orgImpl) AuthenticateMember(
 		return false, nil
 	}
 	return ok, err
-}
-
-// storeImpl is the Store implementation backed by gorm and the identity directory.
-type storeImpl struct {
-	db         *gorm.DB
-	hive       hive.Hive
-	dir        identity.Directory
-	pearDomain string
-	everyone   Org
-}
-
-var _ Store = &storeImpl{}
-
-// NewStore creates a Store that manages multiple orgs on a pear instance.
-func NewStore(
-	db *gorm.DB,
-	hve hive.Hive,
-	dir identity.Directory,
-	pearDomain string,
-) (Store, error) {
-	if err := db.AutoMigrate(&organization{}, &member{}, &spentToken{}); err != nil {
-		return nil, err
-	}
-	return &storeImpl{
-		db:         db,
-		hive:       hve,
-		dir:        dir,
-		pearDomain: pearDomain,
-		everyone:   NewEveryoneOrg(),
-	}, nil
-}
-
-func (s *storeImpl) orgFromModel(org *organization) (*orgImpl, error) {
-	signingSecret, err := base64.StdEncoding.DecodeString(org.SigningSecret)
-	if err != nil {
-		return nil, err
-	}
-	return &orgImpl{
-		orgID:         org.ID,
-		hive:          s.hive,
-		db:            s.db,
-		signingSecret: signingSecret,
-	}, nil
-}
-
-// GetOrg returns the org with the given ID.
-func (s *storeImpl) GetOrg(ctx context.Context, orgID string) (Org, error) {
-	var org organization
-	if err := s.db.WithContext(ctx).Where("id = ?", orgID).First(&org).Error; err != nil {
-		return nil, ErrOrgNotFound
-	}
-	return s.orgFromModel(&org)
-}
-
-// GetOrgForDID returns the org the given DID belongs to.
-// First checks the member table. If the DID isn't in any org, checks if
-// it's managed by our hive. Otherwise returns the everyone org for external DIDs.
-func (s *storeImpl) GetOrgForDID(ctx context.Context, did syntax.DID) (Org, error) {
-	var m member
-	if err := s.db.WithContext(ctx).Where("member = ?", did.String()).First(&m).Error; err == nil {
-		return s.GetOrg(ctx, m.OrgID)
-	}
-
-	id, err := s.dir.LookupDID(ctx, did)
-	if err != nil {
-		return s.everyone, nil
-	}
-
-	svc, hasHabitat := id.Services["habitat"]
-	if hasHabitat && svc.URL == "https://"+s.pearDomain {
-		return nil, ErrMemberNotFound
-	}
-
-	return s.everyone, nil
-}
-
-// AuthenticateMember authenticates a member by handle across all orgs.
-func (s *storeImpl) AuthenticateMember(
-	ctx context.Context,
-	handle string,
-	password string,
-) (bool, error) {
-	id, err := s.hive.LookupHandle(ctx, syntax.Handle(handle))
-	if err != nil {
-		return false, nil
-	}
-
-	org, err := s.GetOrgForDID(ctx, id.DID)
-	if err != nil {
-		return false, nil
-	}
-
-	return org.AuthenticateMember(ctx, handle, password)
-}
-
-// CreateOrg creates a new org with a bootstrap admin member and returns the generated org ID and the admin identity.
-func (s *storeImpl) CreateOrg(
-	ctx context.Context,
-	name string,
-	adminHandle string,
-	adminPassword string,
-) (string, *identity.Identity, error) {
-	// Generate a random org ID
-	orgBytes := make([]byte, 16)
-	if _, err := rand.Read(orgBytes); err != nil {
-		return "", nil, err
-	}
-	orgID := fmt.Sprintf("%x", orgBytes)
-
-	// Generate signing secret
-	secret := make([]byte, 32)
-	if _, err := rand.Read(secret); err != nil {
-		return "", nil, err
-	}
-	signingSecret := base64.StdEncoding.EncodeToString(secret)
-
-	// Hash the admin password
-	passwordHash, err := hashPassword(adminPassword)
-	if err != nil {
-		return "", nil, err
-	}
-
-	// Mint identity for the admin
-	id, persistIdent, err := s.hive.MintIdentity(adminHandle)
-	if err != nil {
-		return "", nil, err
-	}
-
-	// Persist org, identity, and admin member in a single transaction
-	err = s.db.WithContext(ctx).Transaction(func(tx *gorm.DB) error {
-		if err := tx.Create(&organization{
-			ID:            orgID,
-			Name:          name,
-			LoginMethod:   "password", // new orgs default to password auth
-			SigningSecret: signingSecret,
-			CreatedAt:     time.Now(),
-		}).Error; err != nil {
-			if isDuplicateError(err) {
-				return ErrOrgAlreadyExists
-			}
-			return err
-		}
-		if err := persistIdent(tx); err != nil {
-			return err
-		}
-		return tx.Create(&member{
-			OrgID:        orgID,
-			Member:       id.DID.String(),
-			Role:         string(Admin),
-			PasswordHash: passwordHash,
-			CreatedAt:    time.Now(),
-		}).Error
-	})
-	if err != nil {
-		return "", nil, err
-	}
-
-	return orgID, id, nil
 }
